improve backup mode
Up to now, we did initially a full backup and then just incremental
backups. But in the longer view, this is not a) not very efficient
and b) will cost a lot of storage and c) keep certain userdata
forever in the 2ndsitebackup.

Also we did not cleanup any orphaned files etc. so the cache
directory filled very quickly.

So we switch now to the following mode:

1. do an initial backup
2. do incremental backups afterwards for 60 days
3. every 60 days do a full backup.
4. remove everything older than the second last full backup, so
   older than 120 days.
5. cleanup the various files in the beginning.

This should give us a more stable and efficient 2ndsitebackup +
save some diskspace. On the other hand it requires more bandwidth.