introduce soft_failing_targets
The idea is we have a couple of targets that will cause problems over
and over again. Maybe they are currently not backupable or whatever.
To not always put our hands on if such a target is running, we can
add them to a list of softfailing targets, which won't cause
2ndsitebackup to abort immediately, as we would do otherwise in case
of a failure.